refactor: 테스트 코드에서 ActivityType enum 상수 사용으로 일관성 개선
- 문자열 기반 activity code와 수동 ActivityType 생성 로직을 enum 상수 기반으로 변경
- event publisher 패키지 변경에 따라 import 경로 수정
- 중복된 테스트 설정 제거
