+++ src/test/java/com/eventitta/gamification/service/UserActivityServiceTest.java
@@ -2,7 +2,6 @@ package com.eventitta.gamification.service;
 
 import com.eventitta.gamification.domain.ActivityType;
 import com.eventitta.gamification.domain.UserActivity;
-import com.eventitta.gamification.repository.ActivityTypeRepository;
 import com.eventitta.gamification.repository.UserActivityRepository;
 import com.eventitta.user.domain.Provider;
 import com.eventitta.user.domain.Role;
@@ -17,18 +16,19 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.Optional;
 
+import static com.eventitta.gamification.domain.ActivityType.CREATE_POST;
+import static com.eventitta.gamification.domain.ActivityType.USER_LOGIN;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 class UserActivityServiceTest {
 
     @Mock
     private UserActivityRepository userActivityRepository;
-    @Mock
-    private ActivityTypeRepository activityTypeRepository;
     @Mock
     private UserRepository userRepository;
     @Mock
@@ -48,32 +48,20 @@ class UserActivityServiceTest {
             .build();
     }
 
-    private ActivityType createActivityType(String code, int point) {
-        return ActivityType.builder()
-            .id(1L)
-            .code(code)
-            .name("name")
-            .defaultPoint(point)
-            .build();
-    }
-
     @Test
     @DisplayName("활동 기록 시 사용자 포인트가 증가하고 배지 서비스가 호출된다")
     void givenValidActivity_whenRecordActivity_thenPointsIncreasedAndBadgeServiceCalled() {
         // given
         Long userId = 1L;
-        String code = "CREATE_POST";
         Long targetId = 10L;
         User user = createTestUser();
-        ActivityType type = createActivityType(code, 5);
 
         when(userRepository.findById(userId)).thenReturn(Optional.of(user));
-        when(activityTypeRepository.findByCode(code)).thenReturn(Optional.of(type));
         when(userActivityRepository.save(any(UserActivity.class)))
             .thenAnswer(invocation -> invocation.getArgument(0));
 
         // when
-        userActivityService.recordActivity(userId, code, targetId);
+        userActivityService.recordActivity(userId, USER_LOGIN, targetId);
 
         // then
         assertThat(user.getPoints()).isEqualTo(5);
@@ -86,24 +74,22 @@ class UserActivityServiceTest {
     void givenActivity_whenRevokeActivity_thenPointsDecreasedAndActivityDeleted() {
         // given
         Long userId = 1L;
-        String code = "CREATE_POST";
         Long targetId = 10L;
         User user = createTestUser();
         user.earnPoints(20); // 초기 포인트 설정
-        ActivityType type = createActivityType(code, 5);
+        ActivityType type = ActivityType.DELETE_COMMENT;
 
         when(userRepository.findById(userId)).thenReturn(Optional.of(user));
-        when(activityTypeRepository.findByCode(code)).thenReturn(Optional.of(type));
-        when(userActivityRepository.deleteByUserIdAndActivityType_IdAndTargetId(userId, type.getId(), targetId))
+        when(userActivityRepository.deleteByUserIdAndActivityTypeAndTargetId(userId, type, targetId))
             .thenReturn(1L);
 
         // when
-        userActivityService.revokeActivity(userId, code, targetId);
+        userActivityService.revokeActivity(userId, type, targetId);
 
         // then
         assertThat(user.getPoints()).isEqualTo(15);
         verify(userActivityRepository)
-            .deleteByUserIdAndActivityType_IdAndTargetId(userId, type.getId(), targetId);
+            .deleteByUserIdAndActivityTypeAndTargetId(userId, type, targetId);
     }
 
     @Test
@@ -111,30 +97,12 @@ class UserActivityServiceTest {
     void givenNonExistentUser_whenRecordActivity_thenThrowException() {
         // given
         Long userId = 999L;
-        String code = "CREATE_POST";
         Long targetId = 10L;
 
         when(userRepository.findById(userId)).thenReturn(Optional.empty());
 
         // when & then
-        assertThatThrownBy(() -> userActivityService.recordActivity(userId, code, targetId))
-            .isInstanceOf(RuntimeException.class);
-    }
-
-    @Test
-    @DisplayName("존재하지 않는 활동 타입으로 기록 시 예외가 발생한다")
-    void givenNonExistentActivityType_whenRecordActivity_thenThrowException() {
-        // given
-        Long userId = 1L;
-        String code = "INVALID_CODE";
-        Long targetId = 10L;
-        User user = createTestUser();
-
-        when(userRepository.findById(userId)).thenReturn(Optional.of(user));
-        when(activityTypeRepository.findByCode(code)).thenReturn(Optional.empty());
-
-        // when & then
-        assertThatThrownBy(() -> userActivityService.recordActivity(userId, code, targetId))
+        assertThatThrownBy(() -> userActivityService.recordActivity(userId, CREATE_POST, targetId))
             .isInstanceOf(RuntimeException.class);
     }
 }
